Consolidate ProjectAddScenesModal
