--- app-frontend/src/app/components/projects/projectAddScenesModal/projectAddScenesModal.component.js
@@ -1,15 +1,0 @@
-// Component code
-import projectAddScenesModalTpl from './projectAddScenesModal.html';
-
-const ProjectAddScenesModalComponent = {
-    templateUrl: projectAddScenesModalTpl,
-    bindings: {
-        close: '&',
-        dismiss: '&',
-        modalInstance: '<',
-        resolve: '<'
-    },
-    controller: 'ProjectAddScenesModalController'
-};
-
-export default ProjectAddScenesModalComponent;
